Refactor general utility functions
