+++ archinstall/lib/disk/partitioning_menu.py
@@ -3,6 +3,7 @@ from pathlib import Path
 from typing import override
 
 from archinstall.lib.menu.helpers import Confirmation, Input, Selection
+from archinstall.lib.menu.util import prompt_dir
 from archinstall.lib.models.device import (
 	BtrfsMountOption,
 	DeviceModification,
@@ -22,7 +23,6 @@ from archinstall.tui.ui.result import ResultType
 
 from ..menu.list_manager import ListManager
 from ..output import FormattedOutput
-from ..utils.util import prompt_dir
 from .subvolume_menu import SubvolumeMenu
