+++ archinstall/lib/locale/utils.py
@@ -1,6 +1,7 @@
+from archinstall.lib.utils.util import running_from_host
+
 from ..command import SysCommand
 from ..exceptions import ServiceException, SysCallError
-from ..general import running_from_host
 from ..output import error
