+++ archinstall/lib/packages/util.py
@@ -1,9 +1,8 @@
 from functools import lru_cache
-from pathlib import Path
 
 from archinstall.lib.packages.packages import check_package_upgrade
 
-from .output import debug
+from ..output import debug
 
 
 @lru_cache(maxsize=128)
@@ -20,14 +19,3 @@ def check_version_upgrade() -> str | None:
 	debug(f'Archinstall latest: {upgrade}')
 
 	return upgrade
-
-
-def running_from_host() -> bool:
-	"""
-	Check if running from an installed system.
-
-	Returns True if running from installed system (host mode) for host-to-target install.
-	Returns False if /run/archiso exists (ISO mode).
-	"""
-	is_host = not Path('/run/archiso').exists()
-	return is_host
