+++ archinstall/lib/user/user_menu.py
@@ -7,8 +7,8 @@ from archinstall.tui.ui.menu_item import MenuItem
 from archinstall.tui.ui.result import ResultType
 
 from ..menu.list_manager import ListManager
+from ..menu.util import get_password
 from ..models.users import User
-from ..utils.util import get_password
 
 
 class UserList(ListManager[User]):
